==> 2_Data_Pre_Processing.ipynb <==
import pandas as pd
import numpy as np

<h1>** Quarterly Data Cleaning **<h1>

## Quarterly Data Cleaning
df_qtrly_sales = pd.read_csv('df_qtrly_sales.csv', low_memory=False)
df_qtrly_sales.head(4)

df_qtrly_sales.dtypes

# Removing % from all the columns
# Converting them into numeric, They all are object because of mixed data type.
df_qtrly_sales = df_qtrly_sales.replace(',','',regex = True)
df_qtrly_sales = df_qtrly_sales.replace('%','',regex = True)
df_qtrly_sales.head(5)

df_qtrly_sales.columns


df_qtrly_sales.iloc[:,2:] = df_qtrly_sales.iloc[:,2:].apply(pd.to_numeric)
df_qtrly_sales

# changing the names of the columns.. putting year first and month after
col_names = df_qtrly_sales.columns.tolist()
col_name_new = []
for col in col_names:
    try:
        new_col = str(col.split('_')[1]) + str('_') + str(col.split('_')[0])
    except:
        new_col = col
    #print(col,'New name - ',new_col)
    col_name_new.append(new_col)

print(col_names)
print(col_name_new)

col_names_old = df_qtrly_sales.columns
df_qtrly_sales.columns = col_name_new


# Removing extra years, keeping data only after 2018 Dec 
df_qtrly_sales = df_qtrly_sales[['Tikr', 'KPI', '2018_Dec',
                              '2019_Mar', '2019_Jun', '2019_Sep', '2019_Dec', '2020_Mar', '2020_Jun', '2020_Sep', '2020_Dec',
                              '2021_Mar', '2021_Jun', '2021_Sep']]
df_qtrly_sales.columns = ['Tikr', 'KPI', '2018_Q4',
                              '2019_Q1', '2019_Q2', '2019_Q3', '2019_Q4', '2020_Q1', '2020_Q2', '2020_Q3', '2020_Q4',
                              '2021_Q1', '2021_Q2', '2021_Q3']                
df_qtrly_sales.head(5)

# Pivoting the data.. from KPI column to individual columns with quarter
df = df_qtrly_sales
# Pivotted the pivot
pivot_columns = list(set(df.columns) - set(['KPI','Tikr']))  
df_pivot = pd.pivot_table(df, index=['Tikr'], columns=['KPI'], values=pivot_columns, aggfunc='sum')
# Replacing spaces from the column names and joining the two layer of column names into one.
df_pivot.columns =[str(s2) + str('_') + s1  for (s1,s2) in df_pivot.columns.tolist()]
df_pivot.reset_index(inplace=True)
df_pivot.columns = df_pivot.columns.str.replace('+', '')
df_pivot.columns = df_pivot.columns.str.replace(' ', '_')
# Reordering columns by name so that all similar column get together
df_pivot = df_pivot.reindex(sorted(df_pivot.columns), axis=1)
df_pivot.head(4)

# Plan: 
# Test/Train    : I will use data till 2019 as training and data post 2019 as test. [1250 (Train) + (400 Test)]. This is assuming that the Covid impact has only increased the return % but the overall ranking of returns will remain the same.
# Model         : XgBoost + LGBM + RandomForest (dont have sufficient enough data for Neural networks)
# X variable    : Company Financials till 2019.
# Y variable    : last 2 year Return % (2021 vs 2019)- Minus the return from index (so that some compaies which are not performing at par with index will give negative returns, otherwise all companies will give positive returns because of time)
# Results       : We will be categorizing compaies based on binary classification into 10 buckets.. and ideally all higher ranked compaies should fall in the same bracket on the basis of 2020 returns

# Different features that I am planning to create on top of Quarterly and Annual data
#1. QoQ or YoY Percentage change
#2. Average of all these years
#3. CAGR of Sales
#4. Embeddings for sales and revenue etc.
#5. Company Features - 
#   a. MidCap/SmallCap/BlueChip
#   b. Tenure of a company
#   c. Number of shares
#   d. Market cap in 2019
#   e. Company Category/Industry
#   f. Time series cluster - Slowly increasing + Declining + Sharp Rising + Sales increasing + Profit Increasing 

# Future Plan
# 1. Add Cash Flow and holdings data
# 2. Add BSE data (4000 more companies) data
# 3. Add Volumes data
# 4. Add technical indicators like RSI + RSI 200 + RSI 50 etc.
# 5. Add metrics like Petrol + Commodity rates + Inflation rate + India GDP + GDP growth rate %
# 6. Sector Baseline comparisons
# 7.

df_pivot.to_csv('df_qtrly_sales_flat.csv')
#df_pivot = pd.read_csv('df_qtrly_sales_flat.csv').set_index('Unnamed: 0')

sample_string = 'strings'
sample_string[:-2]

# Feature Creation
# How many different type of variables are there
df_pivot.head(4)
col_name = df_pivot.columns
col_type = [i[:-8] for i in col_name]
print(set(col_type))

# There are 16 type of variables. Lets process each of these variables 1 by 1.

# Variable 1    : Sales
matching = [s for s in df_pivot.columns if "Sales\xa0" in s]        # Getting all the columns containing 'sales/profit etc'
matching.append('Tikr')                                             # Added tikr column
matching = matching[-1:] + matching[:-1]                            # Moving Tikr column from bottom to top of the list
#print(matching)
df_sales = df_pivot[matching]
#print(df_sales.head(4))
df_sales_col = df_sales.columns
df_sales_col = [i.replace('\xa0','') for i in df_sales_col]         # Getting rid of special characters 
print(df_sales_col)
print('Number of cols in df - ',len(df_sales_col))
df_sales.columns = df_sales_col

df_sales.head(4)

df_sales.columns

# disable chained assignments
pd.options.mode.chained_assignment = None 

def percentage_change(col1,col2):
    try:
        out = ((col2 - col1) / col1) * 100
    except:
        out = np.nan
    return out
    #return ((df_sales.loc[:,col2] - df_sales.loc[:,col1])/df_sales.loc[:,col1]) * 100

quarter_cols = [['2019_Q1_QoQ','2018_Q4','2019_Q1'],
             ['2019_Q2_QoQ','2019_Q1','2019_Q2'],
             ['2019_Q3_QoQ','2019_Q2','2019_Q3'],
             ['2019_Q4_QoQ','2019_Q3','2019_Q4'],
             ['2020_Q1_QoQ','2019_Q4','2020_Q1'],
             ['2020_Q2_QoQ','2020_Q1','2020_Q2'],
             ['2020_Q3_QoQ','2020_Q2','2020_Q3'],
             ['2020_Q4_QoQ','2020_Q3','2020_Q4'],
             ['2021_Q1_QoQ','2020_Q4','2021_Q1'],
             ['2021_Q2_QoQ','2021_Q1','2021_Q2'],
             ['2021_Q3_QoQ','2021_Q2','2021_Q3']]

def calc_percentage_diff(df_var,var,col_out,col1,col2):
    col1 = str(var) + str('_') + str(col1)
    col2 = str(var) + str('_') + str(col2)
    col_out = str(var) + str('_') + str(col_out)
    #print(col_out,col1,col2)
    df_var[col_out] = percentage_change(df_var[col1],df_var[col2])
    #print(df_sales[col_out])

var = 'Sales'
#vars = ['Sales','Expenses','','']
for combination in quarter_cols:
    var_out = combination[0]
    var1 = combination[1]
    var2 = combination[2]
    calc_percentage_diff(df_sales,var,var_out,var1,var2)

#Throught this loop we will process multiple colmns at once
# We will clean the variables and also calculate the QoQ for each variable

variables = ['Sales','Revenue','Expenses','Net_Profit','Operating_Profit','EPS_in_Rs','Tax_','OPM_']
   
df_quarter = pd.DataFrame()
for var in variables:
    print(var)
    #matching = [s for s in df_pivot.columns if "Sales\xa0" in s]        # Getting all the columns containing 'sales/profit etc'
    matching = [s for s in df_pivot.columns if var in s]        # Getting all the columns containing 'sales/profit etc'
    matching.append('Tikr')                                             # Added tikr column
    matching = matching[-1:] + matching[:-1]                            # Moving Tikr column from bottom to top of the list
    #print(matching)
    df_var = df_pivot[matching]
    #print(df_sales.head(4))
    df_var_col = df_var.columns
    df_var_col = [i.replace('\xa0','') for i in df_var_col]         # Getting rid of special characters 
    print(df_var_col)
    print('Number of cols in df - ',len(df_var_col))
    df_var.columns = df_var_col

    # Calculating QoQ calculations over all available quarter columns
    # List of output and input columns is in quarter_cols
    # Updating each QoQ column 1 by 1
    for combination in quarter_cols:
        try:    
            var_out = combination[0]
            var1 = combination[1]
            var2 = combination[2]
            calc_percentage_diff(df_var,var,var_out,var1,var2)
        except Exception as e:
            print('Error - ',e)
    df_quarter = pd.concat([df_quarter.reset_index(drop=True), df_var], axis=1)

df_quarter.to_csv('df_quarter_sales_flat.csv')

df_quarter.head(4)

<h1>** Annual Data Cleaning **<h1>

## Quarterly Data Cleaning
df_annual_sales = pd.read_csv('df_annual_sales.csv', low_memory=False)
df_annual_sales.head(4)

df_annual_sales.dtypes

# Removing % from all the columns
# Converting them into numeric, They all are object because of mixed data type.
df_annual_sales = df_annual_sales.replace(',','',regex = True)
df_annual_sales = df_annual_sales.replace('%','',regex = True)
df_annual_sales.head(5)

df_annual_sales.columns

df_annual_sales.iloc[:,2:] = df_annual_sales.iloc[:,2:].apply(pd.to_numeric)
df_annual_sales


# Putting year ahead of month in column names
col_names = df_annual_sales.columns.tolist()
col_name_new = []
for col in col_names:
    try:
        new_col = str(col.split('_')[1]) + str('_') + str(col.split('_')[0])
    except:
        new_col = col
    #print(col,'New name - ',new_col)
    col_name_new.append(new_col)

print(col_names)
print(col_name_new)

col_names_old = df_annual_sales.columns
df_annual_sales.columns = col_name_new


# Different companies declare results in different months.. so the numbers are according to that month.
# In the next step we will take an average of different months in a year.. because this is annual table
df_annual_sales = df_annual_sales[['Tikr', 'KPI', '2003_Mar',  '2004_Mar', '2006_Mar', '2006_Sep',
                               '2007_Mar', '2007_Jun', '2007_Sep',                         '2008_Mar', '2008_Jun', '2008_Sep', '2008_Dec',
                               '2009_Mar', '2009_Jun', '2009_Sep', '2009_Nov', '2009_Dec', '2010_Mar', '2010_Jun', '2010_Sep', '2010_Oct', '2010_Dec', 
                               '2011_Mar', '2011_Jun', '2011_Sep', '2011_Oct', '2011_Dec', '2012_Mar', '2012_Jun', '2012_Sep', '2012_Oct', '2012_Dec', 
                               '2013_Mar', '2013_Jun', '2013_Sep', '2013_Oct', '2013_Dec', '2014_Mar', '2014_Jun', '2014_Sep', '2014_Dec',
                               '2015_Mar', '2015_Jun', '2015_Sep', '2015_Dec',             '2016_Mar', '2016_Jun', '2016_Sep', '2016_Dec', 
                               '2017_Mar', '2017_Jun', '2017_Sep', '2017_Dec',             '2018_Mar', '2018_Jun', '2018_Sep', '2018_Dec', 
                               '2019_Mar', '2019_Jun', '2019_Sep', '2019_Dec',             '2020_Mar', '2020_Jun', '2020_Sep', '2020_Dec', 
                               '2021_Mar', '2021_Jun']]                                  
df_annual_sales.head(5)

df_annual_sales[['2019_Mar', '2019_Jun', '2019_Sep', '2019_Dec', ]]

df_annual_sales[['2019_Mar', '2019_Jun', '2019_Sep', '2019_Dec', ]].mean(axis = 1)

# Getting avg of all the months in a year, so that we can get rid of the extra columns 
# because there can be only 1 column in a year.. so averaging will remove the extra columns

df_annual_sales['2006'] = df_annual_sales[['2006_Mar', '2006_Sep']].mean(axis = 1)
df_annual_sales['2007'] = df_annual_sales[['2007_Mar', '2007_Jun', '2007_Sep']].mean(axis = 1)
df_annual_sales['2008'] = df_annual_sales[['2008_Mar', '2008_Jun', '2008_Sep', '2008_Dec']].mean(axis = 1)
df_annual_sales['2009'] = df_annual_sales[['2009_Mar', '2009_Jun', '2009_Sep', '2009_Nov', '2009_Dec']].mean(axis = 1)
df_annual_sales['2010'] = df_annual_sales[['2010_Mar', '2010_Jun', '2010_Sep', '2010_Oct', '2010_Dec']].mean(axis = 1)
df_annual_sales['2011'] = df_annual_sales[['2011_Mar', '2011_Jun', '2011_Sep', '2011_Oct', '2011_Dec']].mean(axis = 1)
df_annual_sales['2012'] = df_annual_sales[['2012_Mar', '2012_Jun', '2012_Sep', '2012_Oct', '2012_Dec']].mean(axis = 1)
df_annual_sales['2013'] = df_annual_sales[['2013_Mar', '2013_Jun', '2013_Sep', '2013_Oct', '2013_Dec']].mean(axis = 1)
df_annual_sales['2014'] = df_annual_sales[['2014_Mar', '2014_Jun', '2014_Sep', '2014_Dec' ]].mean(axis = 1)
df_annual_sales['2015'] = df_annual_sales[['2015_Mar', '2015_Jun', '2015_Sep', '2015_Dec' ]].mean(axis = 1)
df_annual_sales['2016'] = df_annual_sales[['2016_Mar', '2016_Jun', '2016_Sep', '2016_Dec' ]].mean(axis = 1)
df_annual_sales['2017'] = df_annual_sales[['2017_Mar', '2017_Jun', '2017_Sep', '2017_Dec' ]].mean(axis = 1)
df_annual_sales['2018'] = df_annual_sales[['2018_Mar', '2018_Jun', '2018_Sep', '2018_Dec' ]].mean(axis = 1)
df_annual_sales['2019'] = df_annual_sales[['2019_Mar', '2019_Jun', '2019_Sep', '2019_Dec' ]].mean(axis = 1)
df_annual_sales['2020'] = df_annual_sales[['2020_Mar', '2020_Jun', '2020_Sep', '2020_Dec' ]].mean(axis = 1)
df_annual_sales['2021'] = df_annual_sales[['2021_Mar', '2021_Jun' ]].mean(axis = 1)


# Dropping the extra monthly columns
df_annual_sales.drop(['2003_Mar',  '2004_Mar', '2006_Mar', '2006_Sep',
                               '2007_Mar', '2007_Jun', '2007_Sep',                         '2008_Mar', '2008_Jun', '2008_Sep', '2008_Dec',
                               '2009_Mar', '2009_Jun', '2009_Sep', '2009_Nov', '2009_Dec', '2010_Mar', '2010_Jun', '2010_Sep', '2010_Oct', '2010_Dec', 
                               '2011_Mar', '2011_Jun', '2011_Sep', '2011_Oct', '2011_Dec', '2012_Mar', '2012_Jun', '2012_Sep', '2012_Oct', '2012_Dec', 
                               '2013_Mar', '2013_Jun', '2013_Sep', '2013_Oct', '2013_Dec', '2014_Mar', '2014_Jun', '2014_Sep', '2014_Dec',
                               '2015_Mar', '2015_Jun', '2015_Sep', '2015_Dec',             '2016_Mar', '2016_Jun', '2016_Sep', '2016_Dec', 
                               '2017_Mar', '2017_Jun', '2017_Sep', '2017_Dec',             '2018_Mar', '2018_Jun', '2018_Sep', '2018_Dec', 
                               '2019_Mar', '2019_Jun', '2019_Sep', '2019_Dec',             '2020_Mar', '2020_Jun', '2020_Sep', '2020_Dec', 
                               '2021_Mar', '2021_Jun'],axis=1,inplace=True)

# Reordering Columns in df
df_annual_sales = df_annual_sales[['Tikr','KPI','2010','2011','2012','2013','2014','2015','2016','2017','2018','2019','2020','2021']]
df_annual_sales.head(4)

# Pivoting of Annual Table
df = df_annual_sales
# Pivotted the pivot
pivot_columns = list(set(df.columns) - set(['KPI','Tikr']))  
df_pivot = pd.pivot_table(df, index=['Tikr'], columns=['KPI'], values=pivot_columns, aggfunc='sum')
# Replacing spaces from the column names and joining the two layer of column names into one.
df_pivot.columns =[str(s2) + str('_') + s1  for (s1,s2) in df_pivot.columns.tolist()]
df_pivot.reset_index(inplace=True)
df_pivot.columns = df_pivot.columns.str.replace('+', '')
df_pivot.columns = df_pivot.columns.str.replace(' ', '_')
# Reordering columns by name so that all similar column get together
df_pivot = df_pivot.reindex(sorted(df_pivot.columns), axis=1)
df_pivot.head(4)

df_pivot.to_csv('df_annual_sales_flat.csv')
#df_pivot = pd.read_csv('df_annual_sales_flat.csv').set_index('Unnamed: 0')

# Feature Creation
# How many different type of variables are there
df_pivot.head(4)
col_name = df_pivot.columns
col_type = [i[:-5] for i in col_name]
print(set(col_type))
# There are 16 type of variables. Lets process each of these variables 1 by 1.

# Variable 1    : Sales
matching = [s for s in df_pivot.columns if "Sales\xa0" in s]        # Getting all the columns containing 'sales/profit etc'
matching.append('Tikr')                                             # Added tikr column
matching = matching[-1:] + matching[:-1]                            # Moving Tikr column from bottom to top of the list
#print(matching)
df_sales = df_pivot[matching]
#print(df_sales.head(4))
df_sales_col = df_sales.columns
df_sales_col = [i.replace('\xa0','') for i in df_sales_col]         # Getting rid of special characters 
print(df_sales_col)
print('Number of cols in df - ',len(df_sales_col))
df_sales.columns = df_sales_col

# disable chained assignments
pd.options.mode.chained_assignment = None 

def percentage_change(col1,col2):
    try:
        out = ((col2 - col1) / col1) * 100
    except:
        out = np.nan
    return out
    #return ((df_sales.loc[:,col2] - df_sales.loc[:,col1])/df_sales.loc[:,col1]) * 100

annual_cols =  [['2011_YoY','2010','2011'],
                ['2012_YoY','2011','2012'],
                ['2013_YoY','2012','2013'],
                ['2014_YoY','2013','2014'],
                ['2015_YoY','2014','2015'],
                ['2016_YoY','2015','2016'],
                ['2017_YoY','2016','2017'],
                ['2018_YoY','2017','2018'],
                ['2019_YoY','2018','2019'],
                ['2020_YoY','2019','2020'],
                ['2021_YoY','2021','2021']]


def calc_percentage_diff(df_var,var,col_out,col1,col2):
    col1 = str(var) + str('_') + str(col1)
    col2 = str(var) + str('_') + str(col2)
    col_out = str(var) + str('_') + str(col_out)
    #print(col_out,col1,col2)
    df_var[col_out] = percentage_change(df_var[col1],df_var[col2])
    #print(df_sales[col_out])


#Throught this loop we will process multiple colmns at once
# We will clean the variables and also calculate the QoQ for each variable

variables = ['Sales','Revenue','Expenses','Net_Profit','Operating_Profit','EPS_in_Rs','Tax_','OPM_']
   
df_annual = pd.DataFrame()
for var in variables:
    print(var)
    #matching = [s for s in df_pivot.columns if "Sales\xa0" in s]        # Getting all the columns containing 'sales/profit etc'
    matching = [s for s in df_pivot.columns if var in s]        # Getting all the columns containing 'sales/profit etc'
    matching.append('Tikr')                                             # Added tikr column
    matching = matching[-1:] + matching[:-1]                            # Moving Tikr column from bottom to top of the list
    #print(matching)
    df_var = df_pivot[matching]
    #print(df_sales.head(4))
    df_var_col = df_var.columns
    df_var_col = [i.replace('\xa0','') for i in df_var_col]         # Getting rid of special characters 
    print(df_var_col)
    print('Number of cols in df - ',len(df_var_col))
    df_var.columns = df_var_col

    # Calculating QoQ calculations over all available quarter columns
    # List of output and input columns is in quarter_cols
    # Updating each QoQ column 1 by 1
    for combination in annual_cols:
        try:    
            var_out = combination[0]
            var1 = combination[1]
            var2 = combination[2]
            calc_percentage_diff(df_var,var,var_out,var1,var2)
        except Exception as e:
            print('Error - ',e)
    df_annual = pd.concat([df_annual.reset_index(drop=True), df_var], axis=1)

df_annual.to_csv('df_annual_sales_flat.csv')

df_annual.head(5)
