==> 3_BhavCopy_DontUsse copy.ipynb <==
# At first I though I will download all the data from NSE site.. 
# but there are better APIs which will reduce the workload by a ton. Moving to that.

import pandas as pd
import numpy as np
import requests 
from pynse import *

# I will need three bhav copies. 
# 2 bhavcopies for Training - one for 2019 Jan to 2019 Jan

# I can do one more thing.. I can read prev 5 years of data ( Annual & Quarterly)
# Like 2010 to 2015 and then what was the return in 2016, 
# Likewise 2011 to 2016 and output in 2017, this way my rows will increase.

# One more thing.. simply getting one simple number is not going to be good.. 
# I should use MoM returns (12 values) and then avg them.. 
# that way it will be a good indicator of year long performance.

# Version 1: Train: Train the model for 2019 performance with data from 2010 to 2018
#            Validation: Score the model on performance till 2021, Check if the order is relatively correct or not.

# Version 2: Train: Train the model on 5 year back data for every 5 yr window till 2021.
#            Validate the model against any random rows.

import datetime
from datetime import date
import calendar

#************************* sample Code *************************#
#date  = datetime.datetime.now().date()
#date  =  date(2011, 12, 5)

# start_date  = date.replace(day = 1)
# end_date  = date.replace(day = calendar.monthrange(date.year, date.month)[1])

# print(start_date,end_date)
#**************************************************************#

# Created a function to get the list of all dates
def getting_dates():
    list_all_dates = []
    list_month_date = []
    for var_year in range(2010,2021):
        for var_month in range(1,12):
            #print(var_year,var_month)
            mid_date  =  date(var_year, var_month, 15)
            start_date = mid_date.replace(day = 1)
            end_date  = mid_date.replace(day = calendar.monthrange(mid_date.year, mid_date.month)[1])
            #print(start_date,end_date)
            list_month_date = [start_date,end_date]
            list_all_dates.append(list_month_date)
    return list_all_dates
            #return start_date,end_date

all_dates = getting_dates()
print(str(all_dates[0]))

print('today date',str(all_dates[0][0]))
#getting_dates()

print("We need this date in this format {:%d, %b %Y}".format(all_dates[0][0]))
print("We need this date in this format {:%d-%m-%Y}".format(all_dates[0][0]))
print("Year is - {:%Y}".format(all_dates[0][0]))
print("Month is - {:%b}".format(all_dates[0][0]))
print("filenameis - {:%d%b%Y}".format(all_dates[0][0]).upper())


# Since we need 1s & 28th of every month data.. we need to automate the bhavcopy download

#Here is the link for NSE
actual_url = 'https://www1.nseindia.com/content/historical/EQUITIES/2021/DEC/cm03DEC2021bhav.csv.zip'
print(actual_url)
url = "https://www1.nseindia.com/ArchieveSearch?h_filetype=eqbhav&date={:%d-%m-%Y}&section=EQ".format(all_dates[0][0])
url = "https://www1.nseindia.com/content/historical/EQUITIES/{0}/{1}/cm{2}bhav.csv.zip".format("{:%Y}".format(all_dates[0][0]),"{:%b}".format(all_dates[0][0]).upper(),"{:%d%b%Y}".format(all_dates[0][0]).upper())

print(url)

# Now we can hit the URL but we need to download the zip file unzip it and download the actual bhavcopy

for i in range(7):
    while True:
        try:
            a=requests.get('https://nseindia.com/content/historical/EQUITIES/'+y+'/'+dmonth[m]+'/cm'+d+dmonth[m]+y+'bhav.csv.zip')
        except requests.ConnectionError:
            print('No connection, retrying')
        break
