==> 4_Bhavcopy_Processing.ipynb <==
import pandas as pd
import numpy as np
import re

df_bhavcopy = pd.read_csv('df_bhavcopy.csv').set_index('Unnamed: 0')
df_bhavcopy.head(4)

# df_bhavcopy['year'] = pd.DatetimeIndex(df_bhavcopy['TIMESTAMP']).year
# df_bhavcopy['month'] = pd.DatetimeIndex(df_bhavcopy['TIMESTAMP']).month
df_bhavcopy['Yr_Month'] = pd.DatetimeIndex(df_bhavcopy['TIMESTAMP']).strftime('%Y_%m')
df_bhavcopy['DATE'] = pd.to_datetime(df_bhavcopy['TIMESTAMP'])

df_bhavcopy.head(2)

filepath_ca = 'C:\\Saiyad ADT\\Learnnig\\Sensex\\Screener1\\Data_2021_1121\\Corporate_Actions_BSE_2017_2020.csv'    
df_split = pd.read_csv(filepath_ca, header=0)
df_split.head(3)

# Adjusted Stock price for Split and Bonus
def Adjust_stock_splits(df_filter):
    filepath_ca = 'C:\\Saiyad ADT\\Learnnig\\Sensex\\Screener1\\Data_2021_1121\\Corporate_Actions_BSE_2012_2021.csv'    
    df_split = pd.read_csv(filepath_ca, header=0)
    df_split = df_split.rename(columns = {"Security Name":"Tikr","Ex Date":"Ex_Date"})
    df_split['Category'] = ''
    df_split['Category'][df_split['Purpose'].str.contains('Split')] = 'Split'
    df_split['Category'][df_split['Purpose'].str.contains('Bonus')] = 'Bonus'
    
    df_split['Purpose'] = df_split['Purpose'].replace(" to ",":" ,regex = True)
    df_split['Purpose'] = df_split['Purpose'].str.replace('[a-z]+','')
    df_split['Purpose'] = df_split['Purpose'].str.replace('[A-Z]+','')
    charsToRemove = ["$", ".", "€","/","-"]
    df_split['Purpose'] = df_split['Purpose'].str.replace(r"[{}]+".format("".join([re.escape(x) for x in charsToRemove])),"", regex=True)
    df_split['Purpose'] = df_split['Purpose'].str.strip()
    df_split['Ex_Date'] = pd.to_datetime(df_split['Ex_Date'], format = '%d-%b-%y')
    df_split[['Pre', 'Post']] = df_split['Purpose'].str.split(':', 1, expand=True)
    #df_split['Multiplier'] = pd.to_numeric(df_split['Pre'],errors='coerce')/(pd.to_numeric(df_split['Pre'],errors='coerce') + pd.to_numeric(df_split['Post'],errors='coerce'))
    
    #df_split.loc[df_split['Category'] == , 'difficulty'] = 'Easy'

    df_split['Multiplier'] = np.where(df_split['Category'] == 'Split', 
                                        pd.to_numeric(df_split['Post'],errors='coerce')/pd.to_numeric(df_split['Pre'],errors='coerce'), 
                                        pd.to_numeric(df_split['Pre'],errors='coerce')/(pd.to_numeric(df_split['Pre'],errors='coerce') + pd.to_numeric(df_split['Post'],errors='coerce')))
                        

    #df_split = df_split[df_split['Tikr'] == tikr]
    df_split_adj = pd.merge(df_filter, df_split[['Tikr','Ex_Date','Multiplier']], how='left',left_on=['SYMBOL'], right_on=['Tikr'])
    # Pending - To map the numbers and divide the numbers before Start Date
    #sample = df_split_adj.head(2000)
    for i in (['OPEN','HIGH','LOW','CLOSE']):
        df_split_adj[i] = np.where(df_split_adj['DATE'] < df_split_adj['Ex_Date'],df_split_adj[i]*df_split_adj['Multiplier'],df_split_adj[i])
    return df_split_adj

df_bhav_split_adjusted = Adjust_stock_splits(df_bhavcopy)
df_bhav_split_adjusted.to_csv('df_bhavcopy_split_adjusted.csv')


df_bhav_split_adjusted[df_bhav_split_adjusted['SYMBOL'] == 'WIPRO']

#Getting average of the 4 values
df_bhav_split_adjusted['AVG'] = df_bhav_split_adjusted[['OPEN','HIGH','LOW','CLOSE']].mean(axis=1)

df_bhav_pivot = pd.pivot_table(df_bhav_split_adjusted, index=['SYMBOL', 'SERIES'], columns=['Yr_Month'], values=['AVG'], aggfunc='sum')
# Replacing spaces from the column names and joining the two layer of column names into one.
df_bhav_pivot.columns =[str(s2) + str('_') + s1  for (s1,s2) in df_bhav_pivot.columns.tolist()]
df_bhav_pivot.reset_index(inplace=True)
# Reordering columns by name so that all similar column get together
df_bhav_pivot = df_bhav_pivot.reindex(sorted(df_bhav_pivot.columns), axis=1)
df_bhav_pivot.head(3)

df_bhav_yearly = pd.DataFrame()
for i in range(2012,2022):
    #print(i)
    df_bhav_yearly[i] = df_bhav_pivot.filter(regex=str(i)).mean(axis=1)
df_bhav_yearly['Tikr'] = df_bhav_pivot['SYMBOL']
df_bhav_yearly.head()

df_bhav_yearly[df_bhav_yearly['Tikr'] == 'NIFTY50']

len(df_bhav_yearly.columns) - 2
cols = df_bhav_yearly.columns
# Percentae calclation
def percentage_change(col1,col2):
    try:
        out = ((col2 - col1) / col1) * 100
    except:
        out = np.nan
    return out
    #return ((df_sales.loc[:,col2] - df_sales.loc[:,col1])/df_sales.loc[:,col1]) * 100

df_bhav_delta = pd.DataFrame()

# Looping over columns to get the percentage change YoY
for i in range(len(df_bhav_yearly.columns) - 1):
    #print(cols[i])
    if i > 2:
        df_bhav_delta[cols[i]] = percentage_change(df_bhav_yearly[cols[i-3]],df_bhav_yearly[cols[i]])
df_bhav_delta['Tikr'] = df_bhav_yearly['Tikr']
df_bhav_delta.head(4)


# # Averaging the years month over month changes to get an average return

# df_bhav_delta_monthly = pd.DataFrame()
# for i in range(2015,2022):
#     #print(i)
#     df_bhav_delta_monthly[i] = df_bhav_delta.filter(regex=str(i)).mean(axis=1)

# #df_bhav_delta[''] = df.filter(regex='spike')
# df_bhav_delta_monthly['Tikr'] = df_bhav_delta['Tikr']
# df_bhav_delta_monthly.head(4)
df_bhav_delta.to_csv('df_bhav_delta_monthly.csv')

# Initially we were calculating the MoM returns and averaging them. but that did not give any result in modeling..
# So now we moved to 3 year returns, expecting, this will reduce the noise in the system

# Normalization by subtracting the index growth
#df_bhav_delta[df_bhav_delta['Tikr'] == 'WIPRO']
df_bhav_delta[df_bhav_delta['Tikr'] == 'NIFTY50']
# These numbers are pretty high, was the index really performing this well YoY ?

df_bhav_yearly[df_bhav_yearly['Tikr'] == 'NIFTY50']

df_bhav_delta[df_bhav_delta['Tikr'] == 'RAIN']
