==> 5_Merging_Bhav_Fundamental.ipynb <==
import pandas as pd
import numpy as np

df_qtrly = pd.read_csv('df_qtrly_sales_flat.csv').set_index('Unnamed: 0')
df_annual = pd.read_csv('df_annual_sales_flat.csv').set_index('Unnamed: 0')
df_bhav_return = pd.read_csv('df_bhav_delta_monthly.csv').set_index('Unnamed: 0')

df_qtrly.head(3)

df_annual.head(3)

df_bhav_return.head(3)

df_bhav_return[df_bhav_return['Tikr'] == 'NIFTY50']

# Merging the three dataframes
df_fundamentals = pd.merge(df_annual, df_qtrly, how='left',left_on=['Tikr'], right_on=['Tikr'])
df_merge =  pd.merge(df_fundamentals,df_bhav_return, how='left',left_on=['Tikr'], right_on=['Tikr'])
pd.DataFrame(df_merge.columns).to_clipboard()

# As a next step we need to select the 5 year columns for training and the next 1 year return as y variable

list_2015 = ['Tikr','Sales_2010','Sales_2011','Sales_2012','Sales_2013','Sales_2014',
            'Sales_2011_YoY','Sales_2012_YoY','Sales_2013_YoY','Sales_2014_YoY',
            'Revenue_2010','Revenue_2011','Revenue_2012','Revenue_2013','Revenue_2014',
            'Revenue_2011_YoY','Revenue_2012_YoY','Revenue_2013_YoY','Revenue_2014_YoY',
            'Expenses_2010','Expenses_2011','Expenses_2012','Expenses_2013','Expenses_2014',
            'Expenses_2011_YoY','Expenses_2012_YoY','Expenses_2013_YoY','Expenses_2014_YoY',
            'Net_Profit_2010','Net_Profit_2011','Net_Profit_2012','Net_Profit_2013','Net_Profit_2014',
            'Net_Profit_2011_YoY','Net_Profit_2012_YoY','Net_Profit_2013_YoY','Net_Profit_2014_YoY',
            'Operating_Profit_2010','Operating_Profit_2011','Operating_Profit_2012','Operating_Profit_2013','Operating_Profit_2014',
            'Operating_Profit_2011_YoY','Operating_Profit_2012_YoY','Operating_Profit_2013_YoY','Operating_Profit_2014_YoY',
            'EPS_in_Rs_2010','EPS_in_Rs_2011','EPS_in_Rs_2012','EPS_in_Rs_2013','EPS_in_Rs_2014',
            'EPS_in_Rs_2011_YoY','EPS_in_Rs_2012_YoY','EPS_in_Rs_2013_YoY','EPS_in_Rs_2014_YoY',
            'Tax__2010','Tax__2011','Tax__2012','Tax__2013','Tax__2014',
            'Tax__2011_YoY','Tax__2012_YoY','Tax__2013_YoY','Tax__2014_YoY',
            'OPM__2010','OPM__2011','OPM__2012','OPM__2013','OPM__2014',
            'OPM__2011_YoY','OPM__2012_YoY','OPM__2013_YoY','OPM__2014_YoY','2015']

list_2016 = ['Tikr','Sales_2011','Sales_2012','Sales_2013','Sales_2014','Sales_2015',
            'Sales_2012_YoY','Sales_2013_YoY','Sales_2014_YoY','Sales_2015_YoY',
            'Revenue_2011','Revenue_2012','Revenue_2013','Revenue_2014','Revenue_2015',
            'Revenue_2012_YoY','Revenue_2013_YoY','Revenue_2014_YoY','Revenue_2015_YoY',
            'Expenses_2011','Expenses_2012','Expenses_2013','Expenses_2014','Expenses_2015',
            'Expenses_2012_YoY','Expenses_2013_YoY','Expenses_2014_YoY','Expenses_2015_YoY',
            'Net_Profit_2011','Net_Profit_2012','Net_Profit_2013','Net_Profit_2014','Net_Profit_2015',
            'Net_Profit_2012_YoY','Net_Profit_2013_YoY','Net_Profit_2014_YoY','Net_Profit_2015_YoY',
            'Operating_Profit_2011','Operating_Profit_2012','Operating_Profit_2013','Operating_Profit_2014','Operating_Profit_2015',
            'Operating_Profit_2012_YoY','Operating_Profit_2013_YoY','Operating_Profit_2014_YoY','Operating_Profit_2015_YoY',
            'EPS_in_Rs_2011','EPS_in_Rs_2012','EPS_in_Rs_2013','EPS_in_Rs_2014','EPS_in_Rs_2015',
            'EPS_in_Rs_2012_YoY','EPS_in_Rs_2013_YoY','EPS_in_Rs_2014_YoY','EPS_in_Rs_2015_YoY',
            'Tax__2011','Tax__2012','Tax__2013','Tax__2014','Tax__2015',
            'Tax__2012_YoY','Tax__2013_YoY','Tax__2014_YoY','Tax__2015_YoY',
            'OPM__2011','OPM__2012','OPM__2013','OPM__2014','OPM__2015',
            'OPM__2012_YoY','OPM__2013_YoY','OPM__2014_YoY','OPM__2015_YoY','2016']

list_2017 = ['Tikr','Sales_2012','Sales_2013','Sales_2014','Sales_2015','Sales_2016',
            'Sales_2013_YoY','Sales_2014_YoY','Sales_2015_YoY','Sales_2016_YoY',
            'Revenue_2012','Revenue_2013','Revenue_2014','Revenue_2015','Revenue_2016',
            'Revenue_2013_YoY','Revenue_2014_YoY','Revenue_2015_YoY','Revenue_2016_YoY',
            'Expenses_2012','Expenses_2013','Expenses_2014','Expenses_2015','Expenses_2016',
            'Expenses_2013_YoY','Expenses_2014_YoY','Expenses_2015_YoY','Expenses_2016_YoY',
            'Net_Profit_2012','Net_Profit_2013','Net_Profit_2014','Net_Profit_2015','Net_Profit_2016',
            'Net_Profit_2013_YoY','Net_Profit_2014_YoY','Net_Profit_2015_YoY','Net_Profit_2016_YoY',
            'Operating_Profit_2012','Operating_Profit_2013','Operating_Profit_2014','Operating_Profit_2015','Operating_Profit_2016',
            'Operating_Profit_2013_YoY','Operating_Profit_2014_YoY','Operating_Profit_2015_YoY','Operating_Profit_2016_YoY',
            'EPS_in_Rs_2012','EPS_in_Rs_2013','EPS_in_Rs_2014','EPS_in_Rs_2015','EPS_in_Rs_2016',
            'EPS_in_Rs_2013_YoY','EPS_in_Rs_2014_YoY','EPS_in_Rs_2015_YoY','EPS_in_Rs_2016_YoY',
            'Tax__2012','Tax__2013','Tax__2014','Tax__2015','Tax__2016',
            'Tax__2013_YoY','Tax__2014_YoY','Tax__2015_YoY','Tax__2016_YoY',
            'OPM__2012','OPM__2013','OPM__2014','OPM__2015','OPM__2016',
            'OPM__2013_YoY','OPM__2014_YoY','OPM__2015_YoY','OPM__2016_YoY','2017']
    
list_2018 = ['Tikr','Sales_2013','Sales_2014','Sales_2015','Sales_2016','Sales_2017',
            'Sales_2014_YoY','Sales_2015_YoY','Sales_2016_YoY','Sales_2017_YoY',
            'Revenue_2013','Revenue_2014','Revenue_2015','Revenue_2016','Revenue_2017',
            'Revenue_2014_YoY','Revenue_2015_YoY','Revenue_2016_YoY','Revenue_2017_YoY',
            'Expenses_2013','Expenses_2014','Expenses_2015','Expenses_2016','Expenses_2017',
            'Expenses_2014_YoY','Expenses_2015_YoY','Expenses_2016_YoY','Expenses_2017_YoY',
            'Net_Profit_2013','Net_Profit_2014','Net_Profit_2015','Net_Profit_2016','Net_Profit_2017',
            'Net_Profit_2014_YoY','Net_Profit_2015_YoY','Net_Profit_2016_YoY','Net_Profit_2017_YoY',
            'Operating_Profit_2013','Operating_Profit_2014','Operating_Profit_2015','Operating_Profit_2016','Operating_Profit_2017',
            'Operating_Profit_2014_YoY','Operating_Profit_2015_YoY','Operating_Profit_2016_YoY','Operating_Profit_2017_YoY',
            'EPS_in_Rs_2013','EPS_in_Rs_2014','EPS_in_Rs_2015','EPS_in_Rs_2016','EPS_in_Rs_2017',
            'EPS_in_Rs_2014_YoY','EPS_in_Rs_2015_YoY','EPS_in_Rs_2016_YoY','EPS_in_Rs_2017_YoY',
            'Tax__2013','Tax__2014','Tax__2015','Tax__2016','Tax__2017',
            'Tax__2014_YoY','Tax__2015_YoY','Tax__2016_YoY','Tax__2017_YoY',
            'OPM__2013','OPM__2014','OPM__2015','OPM__2016','OPM__2017',
            'OPM__2014_YoY','OPM__2015_YoY','OPM__2016_YoY','OPM__2017_YoY','2018']


list_2019 = ['Tikr','Sales_2014','Sales_2015','Sales_2016','Sales_2017','Sales_2018',
            'Sales_2015_YoY','Sales_2016_YoY','Sales_2017_YoY','Sales_2018_YoY',
            'Revenue_2014','Revenue_2015','Revenue_2016','Revenue_2017','Revenue_2018',
            'Revenue_2015_YoY','Revenue_2016_YoY','Revenue_2017_YoY','Revenue_2018_YoY',
            'Expenses_2014','Expenses_2015','Expenses_2016','Expenses_2017','Expenses_2018',
            'Expenses_2015_YoY','Expenses_2016_YoY','Expenses_2017_YoY','Expenses_2018_YoY',
            'Net_Profit_2014','Net_Profit_2015','Net_Profit_2016','Net_Profit_2017','Net_Profit_2018',
            'Net_Profit_2015_YoY','Net_Profit_2016_YoY','Net_Profit_2017_YoY','Net_Profit_2018_YoY',
            'Operating_Profit_2014','Operating_Profit_2015','Operating_Profit_2016','Operating_Profit_2017','Operating_Profit_2018',
            'Operating_Profit_2015_YoY','Operating_Profit_2016_YoY','Operating_Profit_2017_YoY','Operating_Profit_2018_YoY',
            'EPS_in_Rs_2014','EPS_in_Rs_2015','EPS_in_Rs_2016','EPS_in_Rs_2017','EPS_in_Rs_2018',
            'EPS_in_Rs_2015_YoY','EPS_in_Rs_2016_YoY','EPS_in_Rs_2017_YoY','EPS_in_Rs_2018_YoY',
            'Tax__2014','Tax__2015','Tax__2016','Tax__2017','Tax__2018',
            'Tax__2015_YoY','Tax__2016_YoY','Tax__2017_YoY','Tax__2018_YoY',
            'OPM__2014','OPM__2015','OPM__2016','OPM__2017','OPM__2018',
            'OPM__2015_YoY','OPM__2016_YoY','OPM__2017_YoY','OPM__2018_YoY','2019']

list_2020 = ['Tikr','Sales_2015','Sales_2016','Sales_2017','Sales_2018','Sales_2019',
            'Sales_2016_YoY','Sales_2017_YoY','Sales_2018_YoY','Sales_2019_YoY',
            'Revenue_2015','Revenue_2016','Revenue_2017','Revenue_2018','Revenue_2019',
            'Revenue_2016_YoY','Revenue_2017_YoY','Revenue_2018_YoY','Revenue_2019_YoY',
            'Expenses_2015','Expenses_2016','Expenses_2017','Expenses_2018','Expenses_2019',
            'Expenses_2016_YoY','Expenses_2017_YoY','Expenses_2018_YoY','Expenses_2019_YoY',
            'Net_Profit_2015','Net_Profit_2016','Net_Profit_2017','Net_Profit_2018','Net_Profit_2019',
            'Net_Profit_2016_YoY','Net_Profit_2017_YoY','Net_Profit_2018_YoY','Net_Profit_2019_YoY',
            'Operating_Profit_2015','Operating_Profit_2016','Operating_Profit_2017','Operating_Profit_2018','Operating_Profit_2019',
            'Operating_Profit_2016_YoY','Operating_Profit_2017_YoY','Operating_Profit_2018_YoY','Operating_Profit_2019_YoY',
            'EPS_in_Rs_2015','EPS_in_Rs_2016','EPS_in_Rs_2017','EPS_in_Rs_2018','EPS_in_Rs_2019',
            'EPS_in_Rs_2016_YoY','EPS_in_Rs_2017_YoY','EPS_in_Rs_2018_YoY','EPS_in_Rs_2019_YoY',
            'Tax__2015','Tax__2016','Tax__2017','Tax__2018','Tax__2019',
            'Tax__2016_YoY','Tax__2017_YoY','Tax__2018_YoY','Tax__2019_YoY',
            'OPM__2015','OPM__2016','OPM__2017','OPM__2018','OPM__2019',
            'OPM__2016_YoY','OPM__2017_YoY','OPM__2018_YoY','OPM__2019_YoY','2020']

list_2021 = ['Tikr','Sales_2016','Sales_2017','Sales_2018','Sales_2019','Sales_2020',
            'Sales_2017_YoY','Sales_2018_YoY','Sales_2019_YoY','Sales_2020_YoY',
            'Revenue_2016','Revenue_2017','Revenue_2018','Revenue_2019','Revenue_2020',
            'Revenue_2017_YoY','Revenue_2018_YoY','Revenue_2019_YoY','Revenue_2020_YoY',
            'Expenses_2016','Expenses_2017','Expenses_2018','Expenses_2019','Expenses_2020',
            'Expenses_2017_YoY','Expenses_2018_YoY','Expenses_2019_YoY','Expenses_2020_YoY',
            'Net_Profit_2016','Net_Profit_2017','Net_Profit_2018','Net_Profit_2019','Net_Profit_2020',
            'Net_Profit_2017_YoY','Net_Profit_2018_YoY','Net_Profit_2019_YoY','Net_Profit_2020_YoY',
            'Operating_Profit_2016','Operating_Profit_2017','Operating_Profit_2018','Operating_Profit_2019','Operating_Profit_2020',
            'Operating_Profit_2017_YoY','Operating_Profit_2018_YoY','Operating_Profit_2019_YoY','Operating_Profit_2020_YoY',
            'EPS_in_Rs_2016','EPS_in_Rs_2017','EPS_in_Rs_2018','EPS_in_Rs_2019','EPS_in_Rs_2020',
            'EPS_in_Rs_2017_YoY','EPS_in_Rs_2018_YoY','EPS_in_Rs_2019_YoY','EPS_in_Rs_2020_YoY',
            'Tax__2016','Tax__2017','Tax__2018','Tax__2019','Tax__2020',
            'Tax__2017_YoY','Tax__2018_YoY','Tax__2019_YoY','Tax__2020_YoY',
            'OPM__2016','OPM__2017','OPM__2018','OPM__2019','OPM__2020',
            'OPM__2017_YoY','OPM__2018_YoY','OPM__2019_YoY','OPM__2020_YoY','2021']

df_2015 = df_merge[list_2015]
df_2015['Year'] = '2015'
df_2016 = df_merge[list_2016]
df_2016['Year'] = '2016'
df_2017 = df_merge[list_2017]
df_2017['Year'] = '2017'
df_2018 = df_merge[list_2018]
df_2018['Year'] = '2018'
df_2019 = df_merge[list_2019]
df_2019['Year'] = '2019'
df_2020 = df_merge[list_2020]
df_2020['Year'] = '2020'
df_2021 = df_merge[list_2021]
df_2021['Year'] = '2021'

col_names = ['Tikr','Sales_yr_1','Sales_yr_2','Sales_yr_3','Sales_yr_4','Sales_yr_5',
            'Sales_yr_2_YoY','Sales_yr_3_YoY','Sales_yr_4_YoY','Sales_yr_5_YoY',
            'Revenue_yr_1','Revenue_yr_2','Revenue_yr_3','Revenue_yr_4','Revenue_yr_5',
            'Revenue_yr_2_YoY','Revenue_yr_3_YoY','Revenue_yr_4_YoY','Revenue_yr_5_YoY',
            'Expenses_yr_1','Expenses_yr_2','Expenses_yr_3','Expenses_yr_4','Expenses_yr_5',
            'Expenses_yr_2_YoY','Expenses_yr_3_YoY','Expenses_yr_4_YoY','Expenses_yr_5_YoY',
            'Net_Profit_yr_1','Net_Profit_yr_2','Net_Profit_yr_3','Net_Profit_yr_4','Net_Profit_yr_5',
            'Net_Profit_yr_2_YoY','Net_Profit_yr_3_YoY','Net_Profit_yr_4_YoY','Net_Profit_yr_5_YoY',
            'Operating_Profit_yr_1','Operating_Profit_yr_2','Operating_Profit_yr_3','Operating_Profit_yr_4','Operating_Profit_yr_5',
            'Operating_Profit_yr_2_YoY','Operating_Profit_yr_3_YoY','Operating_Profit_yr_4_YoY','Operating_Profit_yr_5_YoY',
            'EPS_in_Rs_yr_1','EPS_in_Rs_yr_2','EPS_in_Rs_yr_3','EPS_in_Rs_yr_4','EPS_in_Rs_yr_5',
            'EPS_in_Rs_yr_2_YoY','EPS_in_Rs_yr_3_YoY','EPS_in_Rs_yr_4_YoY','EPS_in_Rs_yr_5_YoY',
            'Tax__yr_1','Tax__yr_2','Tax__yr_3','Tax__yr_4','Tax__yr_5',
            'Tax__yr_2_YoY','Tax__yr_3_YoY','Tax__yr_4_YoY','Tax__yr_5_YoY',
            'OPM__yr_1','OPM__yr_2','OPM__yr_3','OPM__yr_4','OPM__yr_5',
            'OPM__yr_2_YoY','OPM__yr_3_YoY','OPM__yr_4_YoY','OPM__yr_5_YoY','Return','Year']

df_2015.columns = col_names
df_2016.columns = col_names
df_2017.columns = col_names
df_2018.columns = col_names
df_2019.columns = col_names
df_2020.columns = col_names
df_2021.columns = col_names

# df_append = pd.concat([df_2015.reset_index(drop=True), df_2016.reset_index(drop=True), df_2017.reset_index(drop=True),
#                        df_2018.reset_index(drop=True), df_2019.reset_index(drop=True), df_2020.reset_index(drop=True)], axis=1)

df_append = pd.concat([df_2015, df_2016, df_2017, df_2018, df_2019, df_2020, df_2021], ignore_index=True)
#df_2020, Removed 2020 from the data purposefully in order to remove the bias that gets added due to Covid

df_append.head()
#df_append.shape

# Getting rid of revenue columns.. since it is equivalent to sales for Financial Companies
df_append['Sales_yr_1'] = df_append[['Sales_yr_1','Revenue_yr_1']].sum(axis=1)
df_append['Sales_yr_2'] = df_append[['Sales_yr_2','Revenue_yr_2']].sum(axis=1)
df_append['Sales_yr_3'] = df_append[['Sales_yr_3','Revenue_yr_3']].sum(axis=1)
df_append['Sales_yr_4'] = df_append[['Sales_yr_4','Revenue_yr_4']].sum(axis=1)
df_append['Sales_yr_5'] = df_append[['Sales_yr_5','Revenue_yr_5']].sum(axis=1)

df_append['Sales_yr_2_YoY'] = df_append[['Sales_yr_2_YoY','Revenue_yr_2_YoY']].sum(axis=1)
df_append['Sales_yr_3_YoY'] = df_append[['Sales_yr_3_YoY','Revenue_yr_3_YoY']].sum(axis=1)
df_append['Sales_yr_4_YoY'] = df_append[['Sales_yr_4_YoY','Revenue_yr_4_YoY']].sum(axis=1)
df_append['Sales_yr_5_YoY'] = df_append[['Sales_yr_5_YoY','Revenue_yr_5_YoY']].sum(axis=1)

df_append.drop(['Revenue_yr_1','Revenue_yr_2','Revenue_yr_3','Revenue_yr_4','Revenue_yr_5',
            'Revenue_yr_2_YoY','Revenue_yr_3_YoY','Revenue_yr_4_YoY','Revenue_yr_5_YoY',],axis=1,inplace=True)

df_append[df_append['Tikr'] == 'RELIANCE']

# Mapping nifty returns to subtract from 
df_bhav_return[df_bhav_return['Tikr'] == 'NIFTY50']

df_nifty_returns = df_bhav_return[df_bhav_return['Tikr'] == 'NIFTY50']
df_nifty_melt = df_nifty_returns.melt(id_vars=["Tikr"], var_name="Year", value_name="Nifty_Returns")
df_nifty_melt

df_final =  pd.merge(df_append,df_nifty_melt, how='left',left_on=['Year'], right_on=['Year'])
df_final.head()

df_final['Return_adj'] = df_final['Return'] - df_final['Nifty_Returns']
df_final.head()

df_final.to_csv('df_train_data.csv')
